+++ TEST/Injector/Graph.cs
@@ -18,7 +18,7 @@ namespace Solti.Utils.DI.Injector.Graph.Tests
     {
         private static AbstractServiceReference[] Validate(Injector injector) 
         {
-            Config.Value.StrictDI = false;
+            Config.Value.Injector.StrictDI = false;
 
             AbstractServiceReference
                 svc1, svc2, svc3, svc4;